+++ .dotnet/src/Custom/RealtimeConversation/Internal/InternalRealtimeClientEventResponseCreateResponse.cs
@@ -8,13 +8,13 @@ using System.Linq;
 namespace OpenAI.RealtimeConversation;
 
 [Experimental("OPENAI002")]
-[CodeGenModel("RealtimeClientEventResponseCreateResponse")]
-internal partial class InternalRealtimeClientEventResponseCreateResponse
+[CodeGenModel("RealtimeResponseOptions")]
+internal partial class InternalRealtimeResponseOptions
 {
     [CodeGenMember("ToolChoice")]
     public BinaryData ToolChoice { get; set; }
 
-    public static InternalRealtimeClientEventResponseCreateResponse FromSessionOptions(
+    public static InternalRealtimeResponseOptions FromSessionOptions(
         ConversationSessionOptions sessionOptions)
     {
         Argument.AssertNotNull(sessionOptions, nameof(sessionOptions));
@@ -28,17 +28,14 @@ internal partial class InternalRealtimeClientEventResponseCreateResponse
             : null;
         IList<InternalRealtimeRequestSessionModality> internalModalities
             = sessionOptions.ContentModalities.ToInternalModalities();
-        IList<string> rawModalities = internalModalities.Count > 0
-            ? internalModalities.Select(modality => modality.ToString()).ToList()
-            : new ChangeTrackingList<string>();
         BinaryData toolChoice = Optional.IsDefined(sessionOptions.ToolChoice)
             ? ModelReaderWriter.Write(sessionOptions.ToolChoice)
             : null;
-        InternalRealtimeClientEventResponseCreateResponse internalOptions = new(
-            modalities: rawModalities,
+        InternalRealtimeResponseOptions internalOptions = new(
+            modalities: internalModalities,
             instructions: sessionOptions.Instructions,
-            voice: sessionOptions.Voice?.ToString(),
-            outputAudioFormat: sessionOptions.OutputAudioFormat?.ToString(),
+            voice: sessionOptions.Voice,
+            outputAudioFormat: sessionOptions.OutputAudioFormat,
             tools: sessionOptions.Tools,
             toolChoice: toolChoice,
             temperature: sessionOptions.Temperature,